+++ Microsoft.Azure.Cosmos/src/Linq/BuiltinFunctions/TypeCheckFunctions.cs
@@ -18,6 +18,22 @@ namespace Microsoft.Azure.Cosmos.Linq
         {
             TypeCheckFunctionsDefinitions = new Dictionary<string, BuiltinFunctionVisitor>();
 
+            TypeCheckFunctionsDefinitions.Add("IsArray",
+                new SqlBuiltinFunctionVisitor("IS_ARRAY",
+                    true,
+                    new List<Type[]>()
+                    {
+                        new Type[]{typeof(object)},
+                    }));
+
+            TypeCheckFunctionsDefinitions.Add("IsBool",
+                new SqlBuiltinFunctionVisitor("IS_BOOL",
+                    true,
+                    new List<Type[]>()
+                    {
+                        new Type[]{typeof(object)},
+                    }));
+
             TypeCheckFunctionsDefinitions.Add("IsDefined",
                 new SqlBuiltinFunctionVisitor("IS_DEFINED",
                     true,
@@ -34,8 +50,32 @@ namespace Microsoft.Azure.Cosmos.Linq
                         new Type[]{typeof(object)},
                     }));
 
+            TypeCheckFunctionsDefinitions.Add("IsNumber",
+                new SqlBuiltinFunctionVisitor("IS_NUMBER",
+                    true,
+                    new List<Type[]>()
+                    {
+                        new Type[]{typeof(object)},
+                    }));
+
+            TypeCheckFunctionsDefinitions.Add("IsObject",
+                new SqlBuiltinFunctionVisitor("IS_OBJECT",
+                    true,
+                    new List<Type[]>()
+                    {
+                        new Type[]{typeof(object)},
+                    }));
+
             TypeCheckFunctionsDefinitions.Add("IsPrimitive",
                 new SqlBuiltinFunctionVisitor("IS_PRIMITIVE",
+                    true,
+                    new List<Type[]>()
+                    {
+                        new Type[]{typeof(object)},
+                    }));
+
+            TypeCheckFunctionsDefinitions.Add("IsString",
+                new SqlBuiltinFunctionVisitor("IS_STRING",
                     true,
                     new List<Type[]>()
                     {